Daxa 3.6: fix missing indirect-dispatch barrier on the physics DispatchBuffer
RigidBodyDispatcher writes the per-pass indirect dispatch counts into
DispatchBuffer, but the morton-code, radix-sort (histogram + sort), LBVH
hierarchy / bounding-box build / convert, and reordering task heads declared
that buffer as COMPUTE_SHADER_READ only. The task graph therefore did not treat
it as an indirect-command source and inserted no barrier between the
dispatcher's write and the dispatch_indirect reads.

The resulting race let the indirect dispatches read stale counts, so the sort
and reorder ran over wrong ranges and a growing fraction of bodies ended up
holding the static floor body's transform and +/-50 AABB. Those huge leaf AABBs
collapsed the LBVH broad phase to O(N^2), burying the narrow phase and solver:
past a few hundred bodies the cubes froze / flew off / vanished and the
simulation crashed.

Fix: add COMPUTE_SHADER_READ_INDIRECT_COMMAND_READ and use it for the
dispatch_buffer attachment in every consumer task head (include/shared.inl) so
Daxa inserts the correct write->indirect-read barrier. Also harden the LBVH
broad-phase traversal stack (named LBVH_TRAVERSAL_STACK_SIZE, bounds-guarded
push). Includes accompanying Daxa 3.6 migration refinements.

+++ include/camera.h
@@ -8,8 +8,8 @@ const glm::vec3 INIT_CAMERA_POS = {0, 1, -4.5};
 const glm::vec3 INIT_CAMERA_UP = {0, 1, 0};
 
 const float INIT_CAMERA_FOV = 45.0f;
-const float INIT_CAMERA_WIDTH = 800.0f;
-const float INIT_CAMERA_HEIGHT = 600.0f;
+const unsigned int INIT_CAMERA_WIDTH = 800;
+const unsigned int INIT_CAMERA_HEIGHT = 600;
 const float INIT_CAMERA_NEAR = 0.001f;
 const float INIT_CAMERA_FAR = 1000.0f;
 const float CAMERA_SPEED = 0.1f;
@@ -145,7 +145,7 @@ FORCE_INLINE const float& camera_get_fov(const Camera& cam) {
 }
 
 FORCE_INLINE const float camera_get_aspect(const Camera& cam) {
-    return cam.width / cam.height;
+    return static_cast<float>(cam.width) / static_cast<float>(cam.height);
 }
 
 FORCE_INLINE const unsigned int& camera_get_width(const Camera& cam) {
